refactor out detignore and v1file utils
